Refactor /organisms/[entity] to use generated types
Replaced imports of AccessionEvent and CollectionEvent from '@/queries/specimen' with '@/generated/types' in page.tsx and Collecting.tsx for consistency and to use generated GraphQL types.
